Add 'ui.dialog' subpackage
