+++ src/edu/colorado/csdms/wmt/client/ui/dialog/NewUserDialogBox.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package edu.colorado.csdms.wmt.client.ui.widgets;
+package edu.colorado.csdms.wmt.client.ui.dialog;
 
 import com.google.gwt.dom.client.Style.Unit;
 import com.google.gwt.user.client.ui.DialogBox;
@@ -29,79 +29,55 @@ import com.google.gwt.user.client.ui.HTML;
 import com.google.gwt.user.client.ui.HasHorizontalAlignment;
 import com.google.gwt.user.client.ui.VerticalPanel;
 
+import edu.colorado.csdms.wmt.client.Constants;
 import edu.colorado.csdms.wmt.client.ui.panel.ChoicePanel;
-import edu.colorado.csdms.wmt.client.ui.panel.DroplistPanel;
 import edu.colorado.csdms.wmt.client.ui.panel.FieldPanel;
 
 /**
- * A customized DialogBox with a droplist for choosing the host to run the
- * model, as well as input fields for logging into the host.
+ * A {@link DialogBox} that prompts a user to reenter their password when
+ * creating a new WMT sign in.
  * 
  * @author Mark Piper (mark.piper@colorado.edu)
  */
-public class RunDialogBox extends DialogBox {
+public class NewUserDialogBox extends DialogBox {
 
-  private DroplistPanel hostPanel;
-  private FieldPanel usernamePanel;
   private FieldPanel passwordPanel;
   private ChoicePanel choicePanel;
   
   /**
-   * Makes a new RunDialogBox.
+   * Creates a {@link NewUserDialogBox}.
+   * 
+   * @param data the DataManager object for the WMT session
    */
-  public RunDialogBox() {
+  public NewUserDialogBox() {
 
     super(false); // autohide
     this.setModal(true);
+    this.setText("New User");
     this.setStyleName("wmt-DialogBox");
-    this.setText("Run Model...");
-
-    hostPanel = new DroplistPanel();
-
-    HTML separator = new HTML();
-    separator.setStyleName("wmt-PopupPanelSeparator");
     
-    usernamePanel = new FieldPanel();
+    String msg = "This email address is not registered with WMT."
+        + " If you would like to use this address as your sign in,"
+        + " please reenter your password below and click \"New User\";"
+        + " if not, click \"Cancel\".";
+    HTML msgHtml = new HTML(msg);
+    
     passwordPanel = new FieldPanel(true); // uses PasswordTextBox
-    choicePanel = new ChoicePanel();
+    passwordPanel.getLabel().setText("Reenter password:");
     
-    hostPanel.getLabel().setText("Host:");
-    usernamePanel.getLabel().setText("Username:");
-    passwordPanel.getLabel().setText("Password:");
-    choicePanel.getOkButton().setHTML("<i class='fa fa-play'></i> Run");
-
-    VerticalPanel panel = new VerticalPanel();
-    panel.setWidth("100%");
-    panel.setHorizontalAlignment(HasHorizontalAlignment.ALIGN_CENTER);
-    panel.setSpacing(10); // px
-    panel.add(hostPanel);
-    panel.add(separator);
+    choicePanel = new ChoicePanel();
+    choicePanel.getOkButton().setHTML(Constants.FA_USER + "New User");
     
     VerticalPanel contents = new VerticalPanel();
-    contents.setHorizontalAlignment(HasHorizontalAlignment.ALIGN_RIGHT);
-    contents.getElement().getStyle().setMargin(5, Unit.PX);
-    contents.add(panel);
-    contents.add(usernamePanel);
+    contents.setHorizontalAlignment(HasHorizontalAlignment.ALIGN_CENTER);
+    contents.setWidth("30em");
+    contents.setSpacing(5); // px
+    contents.getElement().getStyle().setPaddingTop(5.0, Unit.PX);
+    contents.add(msgHtml);
     contents.add(passwordPanel);
     contents.add(choicePanel);
 
     this.setWidget(contents);
-  }
-
-  public DroplistPanel getHostPanel() {
-    return hostPanel;
-  }
-
-  public void setHostPanel(DroplistPanel hostPanel) {
-    this.hostPanel = hostPanel;
-  }
-
-  public FieldPanel getUsernamePanel() {
-    return usernamePanel;
-  }
-
-  public void setUsernamePanel(FieldPanel usernamePanel) {
-    this.usernamePanel = usernamePanel;
   }
 
   public FieldPanel getPasswordPanel() {
