+++ src/edu/colorado/csdms/wmt/client/ui/dialog/RunDialogBox.java
@@ -21,53 +21,95 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package edu.colorado.csdms.wmt.client.ui.widgets;
+package edu.colorado.csdms.wmt.client.ui.dialog;
 
+import com.google.gwt.dom.client.Style.Unit;
 import com.google.gwt.user.client.ui.DialogBox;
+import com.google.gwt.user.client.ui.HTML;
 import com.google.gwt.user.client.ui.HasHorizontalAlignment;
 import com.google.gwt.user.client.ui.VerticalPanel;
 
 import edu.colorado.csdms.wmt.client.ui.panel.ChoicePanel;
 import edu.colorado.csdms.wmt.client.ui.panel.DroplistPanel;
+import edu.colorado.csdms.wmt.client.ui.panel.FieldPanel;
 
 /**
- * A customized DialogBox with a {@link DroplistPanel} for choosing an item
- * and a {@link ChoicePanel} displaying "OK" and "Cancel" buttons.
+ * A customized DialogBox with a droplist for choosing the host to run the
+ * model, as well as input fields for logging into the host.
  * 
  * @author Mark Piper (mark.piper@colorado.edu)
  */
-public class DroplistDialogBox extends DialogBox {
+public class RunDialogBox extends DialogBox {
 
-  private DroplistPanel itemPanel;
+  private DroplistPanel hostPanel;
+  private FieldPanel usernamePanel;
+  private FieldPanel passwordPanel;
   private ChoicePanel choicePanel;
   
   /**
-   * Makes a new DroplistDialogBox with default settings in the
-   * {@link DroplistPanel} and {@link ChoicePanel}.
+   * Makes a new RunDialogBox.
    */
-  public DroplistDialogBox() {
+  public RunDialogBox() {
 
     super(false); // autohide
     this.setModal(true);
     this.setStyleName("wmt-DialogBox");
+    this.setText("Run Model...");
 
-    itemPanel = new DroplistPanel();
+    hostPanel = new DroplistPanel();
+
+    HTML separator = new HTML();
+    separator.setStyleName("wmt-PopupPanelSeparator");
+    
+    usernamePanel = new FieldPanel();
+    passwordPanel = new FieldPanel(true); // uses PasswordTextBox
     choicePanel = new ChoicePanel();
+    
+    hostPanel.getLabel().setText("Host:");
+    usernamePanel.getLabel().setText("Username:");
+    passwordPanel.getLabel().setText("Password:");
+    choicePanel.getOkButton().setHTML("<i class='fa fa-play'></i> Run");
 
+    VerticalPanel panel = new VerticalPanel();
+    panel.setWidth("100%");
+    panel.setHorizontalAlignment(HasHorizontalAlignment.ALIGN_CENTER);
+    panel.setSpacing(10); // px
+    panel.add(hostPanel);
+    panel.add(separator);
+    
     VerticalPanel contents = new VerticalPanel();
     contents.setHorizontalAlignment(HasHorizontalAlignment.ALIGN_RIGHT);
-    contents.add(itemPanel);
+    contents.getElement().getStyle().setMargin(5, Unit.PX);
+    contents.add(panel);
+    contents.add(usernamePanel);
+    contents.add(passwordPanel);
     contents.add(choicePanel);
 
     this.setWidget(contents);
   }
 
-  public DroplistPanel getDroplistPanel() {
-    return itemPanel;
+  public DroplistPanel getHostPanel() {
+    return hostPanel;
   }
 
-  public void setDroplistPanel(DroplistPanel itemPanel) {
-    this.itemPanel = itemPanel;
+  public void setHostPanel(DroplistPanel hostPanel) {
+    this.hostPanel = hostPanel;
+  }
+
+  public FieldPanel getUsernamePanel() {
+    return usernamePanel;
+  }
+
+  public void setUsernamePanel(FieldPanel usernamePanel) {
+    this.usernamePanel = usernamePanel;
+  }
+
+  public FieldPanel getPasswordPanel() {
+    return passwordPanel;
+  }
+
+  public void setPasswordPanel(FieldPanel passwordPanel) {
+    this.passwordPanel = passwordPanel;
   }
 
   public ChoicePanel getChoicePanel() {
