+++ src/edu/colorado/csdms/wmt/client/ui/dialog/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Dialogs (or {@link com.google.gwt.user.client.ui.DialogBox}es) are modal
+ * panels that float above the WMT interface. They need to be clicked to be
+ * dismissed.
+ * 
+ * @author Mark Piper (mark.piper@colorado.edu)
+ */
+package edu.colorado.csdms.wmt.client.ui.dialog;
