+++ src/edu/colorado/csdms/wmt/client/ui/handler/AuthenticationHandler.java
@@ -31,7 +31,7 @@ import com.google.gwt.event.dom.client.KeyDownEvent;
 import edu.colorado.csdms.wmt.client.Constants;
 import edu.colorado.csdms.wmt.client.control.DataManager;
 import edu.colorado.csdms.wmt.client.control.DataTransfer;
-import edu.colorado.csdms.wmt.client.ui.widgets.QuestionDialogBox;
+import edu.colorado.csdms.wmt.client.ui.dialog.QuestionDialogBox;
 
 /**
  * Handles login and logout events in the WMT client.
