+++ src/edu/colorado/csdms/wmt/client/ui/handler/DeleteModelHandler.java
@@ -29,7 +29,7 @@ import com.google.gwt.event.dom.client.ClickHandler;
 
 import edu.colorado.csdms.wmt.client.control.DataManager;
 import edu.colorado.csdms.wmt.client.control.DataTransfer;
-import edu.colorado.csdms.wmt.client.ui.widgets.DroplistDialogBox;
+import edu.colorado.csdms.wmt.client.ui.dialog.DroplistDialogBox;
 
 /**
  * Handles click on the "Delete" button in the dialog that appears when the
