+++ src/edu/colorado/csdms/wmt/client/ui/handler/ParameterActionPanelResetHandler.java
@@ -28,7 +28,7 @@ import com.google.gwt.event.dom.client.ClickHandler;
 
 import edu.colorado.csdms.wmt.client.Constants;
 import edu.colorado.csdms.wmt.client.control.DataManager;
-import edu.colorado.csdms.wmt.client.ui.widgets.QuestionDialogBox;
+import edu.colorado.csdms.wmt.client.ui.dialog.QuestionDialogBox;
 
 /**
  * Handles a click on the "Reset" button in the ParameterActionPanel. Calls
