+++ src/edu/colorado/csdms/wmt/client/ui/handler/RunModelHandler.java
@@ -29,7 +29,7 @@ import com.google.gwt.event.dom.client.ClickHandler;
 
 import edu.colorado.csdms.wmt.client.control.DataManager;
 import edu.colorado.csdms.wmt.client.control.DataTransfer;
-import edu.colorado.csdms.wmt.client.ui.widgets.RunDialogBox;
+import edu.colorado.csdms.wmt.client.ui.dialog.RunDialogBox;
 
 /**
  * Handles click on the "Run" button in the dialog that appears when the
