+++ src/edu/colorado/csdms/wmt/client/ui/handler/SaveModelHandler.java
@@ -30,7 +30,7 @@ import com.google.gwt.user.client.Window;
 import edu.colorado.csdms.wmt.client.Constants;
 import edu.colorado.csdms.wmt.client.control.DataManager;
 import edu.colorado.csdms.wmt.client.control.DataTransfer;
-import edu.colorado.csdms.wmt.client.ui.widgets.SaveDialogBox;
+import edu.colorado.csdms.wmt.client.ui.dialog.SaveDialogBox;
 
 /**
  * Handles click on the "OK" button in the save dialog that appears when the
