+++ src/edu/colorado/csdms/wmt/client/ui/widgets/ValueCell.java
@@ -42,6 +42,7 @@ import com.google.gwt.user.client.ui.TextBox;
 import edu.colorado.csdms.wmt.client.control.DataURL;
 import edu.colorado.csdms.wmt.client.data.ParameterJSO;
 import edu.colorado.csdms.wmt.client.ui.ParameterTable;
+import edu.colorado.csdms.wmt.client.ui.dialog.UploadDialogBox;
 
 /**
  * Used to display the value of a parameter in a {@link ParameterTable}, a
